requirements.txt
streamlit
fpdf
PyPDF2
python-docx
